extract error reporting resolve into a standalone file
